fix(ci): fix build command and lint issues
- Fix build command to use zen-flow-controller instead of gc-controller
- Fix G115 integer overflow warnings with nosec comments
- Add nolint comments for cyclomatic complexity in reconciler

+++ pkg/logging/logger.go
@@ -172,7 +172,7 @@ func safeKlogLevel(level int) klog.Level {
 	} else if level > 10 {
 		level = 10
 	}
-	return klog.Level(level)
+	return klog.Level(int32(level)) // #nosec G115 -- level is bounded [0,10]
 }
 
 // Info logs at the specified verbosity level.
